feat(remove-babel): Removed babel.
Separated yargs code and the auditing. This will support testing.
