+++ .eslintignore
@@ -1,3 +1,2 @@
 /bin
 /node_modules
-index.js
