+++ .eslintrc.json
@@ -2,8 +2,5 @@
   "env": {
     "node": true
   },
-  "extends": "airbnb-base",
-  "rules": {
-    "linebreak-style": "off"
-  }
+  "extends": "airbnb-base"
 }
