+++ bin/audit-ci
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../');
+require('../lib/audit-ci');
