+++ package.json
@@ -1,9 +1,9 @@
 {
   "name": "audit-ci",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "Manages running npm audit in CI environments",
   "license": "Apache-2.0",
-  "main": "./index.js",
+  "main": "./lib/audit-ci.js",
   "repository": {
     "type": "git",
     "url": "git@github.com:IBM/audit-ci.git"
@@ -22,25 +22,20 @@
   },
   "files": [
     "bin",
-    "index.js",
+    "lib/*",
     "README.md"
   ],
   "scripts": {
-    "build": "babel src/audit-ci.js | terser --compress --mangle -o index.js",
-    "lint": "eslint -c ./.eslintrc.json src/audit-ci.js",
-    "prepare": "npm run build"
+    "lint": "eslint -c ./.eslintrc.json lib/*",
+    "test": "npm run lint"
   },
   "dependencies": {
-    "npm-audit-report": "^1.3.1",
-    "yargs": "^12.0.2"
+    "npm-audit-report": "1.3.1",
+    "yargs": "12.0.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.5",
-    "@babel/core": "^7.1.6",
-    "@babel/preset-env": "^7.1.6",
-    "eslint": "^5.9.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "terser": "^3.10.11"
+    "eslint": "5.9.0",
+    "eslint-config-airbnb-base": "13.1.0",
+    "eslint-plugin-import": "2.14.0"
   }
 }
